requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.1
gunicorn==19.6.0